+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 251:
 
 * Clean up CI scripts
 * detect_ssl uses bool
+* launder pointers
 
 --------------------------------------------------------------------------------